requirements.txt
Django==2.0.2
requests==2.9.1
lxml==3.4.4
hiredis==0.2.0
redis==2.10.3
oss2==2.0.5
python-dateutil==2.5.0
beautifulsoup4==4.4.1
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
splinter==0.7.7